mod: #5 main -> mypage 코드 이동
