+++ app/src/main/java/com/sopt/now/MainActivity.kt
@@ -3,9 +3,6 @@ package com.sopt.now
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.sopt.now.databinding.ActivityMainBinding
-import com.sopt.now.LoginActivity.Companion.ID
-import com.sopt.now.LoginActivity.Companion.MBTI
-import com.sopt.now.LoginActivity.Companion.NICKNAME
 
 class MainActivity : AppCompatActivity() {
     lateinit var binding : ActivityMainBinding
@@ -14,17 +11,7 @@ class MainActivity : AppCompatActivity() {
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        initTextViews()
 
     }
 
-    private fun initTextViews() {
-        intent.apply {
-            with(binding) {
-                tvMainNickname.setText(getStringExtra(NICKNAME))
-                tvMainIdContent.setText(getStringExtra(ID))
-                tvMainMbtiContent.setText(getStringExtra(MBTI))
-            }
-        }
-    }
 }
