+++ app/src/main/java/com/sopt/now/MypageFragment.kt
@@ -6,6 +6,9 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.sopt.now.databinding.FragmentMypageBinding
+import com.sopt.now.LoginActivity.Companion.ID
+import com.sopt.now.LoginActivity.Companion.MBTI
+import com.sopt.now.LoginActivity.Companion.NICKNAME
 
 class MypageFragment : Fragment() {
     private var _binding : FragmentMypageBinding? = null
@@ -23,11 +26,21 @@ class MypageFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
+        initTextViews()
     }
 
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
+    }
+
+    private fun initTextViews() {
+        activity?.intent?.apply {
+            with(binding) {
+                tvMainNickname.setText(getStringExtra(NICKNAME))
+                tvMainIdContent.setText(getStringExtra(ID))
+                tvMainMbtiContent.setText(getStringExtra(MBTI))
+            }
+        }
     }
 }
