+++ app/src/main/res/layout/fragment_mypage.xml
@@ -2,13 +2,75 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:padding="70dp"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
+    <ImageView
+        android:id="@+id/iv_main_profile"
+        android:layout_width="200dp"
+        android:layout_height="0dp"
+        android:src="@drawable/ic_launcher_foreground"
+        android:background="#B4E2D4"
+        app:layout_constraintDimensionRatio="1"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"/>
+
+    <Space
+        android:id="@+id/sp_main_spacer"
+        android:layout_width="wrap_content"
+        android:layout_height="30dp"
+        app:layout_constraintTop_toBottomOf="@id/iv_main_profile"
+        app:layout_constraintStart_toStartOf="@id/iv_main_profile"/>
+
     <TextView
+        android:id="@+id/tv_main_nickname"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="mypage"
+        android:textSize="30sp"
+        android:textStyle="bold"
+        app:layout_constraintTop_toBottomOf="@id/sp_main_spacer"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"/>
+        app:layout_constraintEnd_toEndOf="parent"/>
+
+    <TextView
+        android:id="@+id/tv_main_id"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginVertical="40dp"
+        android:text="@string/tv_main_id"
+        android:textSize="30sp"
+        android:textStyle="bold"
+        app:layout_constraintTop_toBottomOf="@id/tv_main_nickname"
+        app:layout_constraintStart_toStartOf="parent"/>
+
+    <TextView
+        android:id="@+id/tv_main_id_content"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:paddingHorizontal="20dp"
+        android:textSize="30sp"
+        app:layout_constraintTop_toTopOf="@id/tv_main_id"
+        app:layout_constraintStart_toEndOf="@id/tv_main_id"/>
+
+    <TextView
+        android:id="@+id/tv_main_mbti"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginVertical="30dp"
+        android:text="@string/tv_main_mbti"
+        android:textSize="30sp"
+        android:textStyle="bold"
+        app:layout_constraintTop_toBottomOf="@id/tv_main_id"
+        app:layout_constraintStart_toStartOf="parent"/>
+
+    <TextView
+        android:id="@+id/tv_main_mbti_content"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:paddingHorizontal="20dp"
+        android:textSize="30sp"
+        app:layout_constraintTop_toTopOf="@id/tv_main_mbti"
+        app:layout_constraintStart_toEndOf="@id/tv_main_mbti"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
